+++ codigo/funGenetico.hpp
@@ -20,7 +20,7 @@ using namespace std::chrono;
 
 Genoma hacer_genetico(  unsigned int generaciones,
                         unsigned int tamanio_poblacion,
-                        unsigned int proba_mutacion,
+                        double proba_mutacion,
                         unsigned int n,
                         unsigned int m,
                         unsigned int total,
@@ -30,7 +30,7 @@ Genoma hacer_genetico(  unsigned int generaciones,
                         function<pair<vector<Genoma>,vector<Genoma> > (vector<Genoma> &poblacion,
                                                 vector<int> &puntajes,
                                                 double fracc_poblacion)> seleccion,
-                        function<void(Genoma &individuo, unsigned int proba_mutacion)> mutacion,
+                        function<void(Genoma &individuo, double proba_mutacion)> mutacion,
                         function<Genoma(Genoma &a, Genoma &b)> crossover );
 
 vector<Genoma> poblacion_inicial(unsigned int tamanio_poblacion);
@@ -45,13 +45,13 @@ pair<vector<Genoma>,vector<Genoma> > seleccion_por_puntaje(vector<Genoma> &pobla
 pair<vector<Genoma>,vector<Genoma> > seleccion_por_cantidad(vector<Genoma> &poblacion,
                                                             vector<int> &puntajes, double fracc_poblacion);
 
-void mutacion_A(Genoma &individuo, unsigned int proba_mutacion);
-void mutacion_B(Genoma &individuo, unsigned int proba_mutacion);
+void mutacion_A(Genoma &individuo, double proba_mutacion);
+void mutacion_B(Genoma &individuo, double proba_mutacion);
 
 vector<Genoma> hacer_crossover( vector<Genoma> &poblacion,
                                 function<void(Genoma &individuo,
-                                                unsigned int proba_mutacion)> mutacion,
-                                unsigned int proba_mutacion,
+                                                double proba_mutacion)> mutacion,
+                                double proba_mutacion,
                                 function<Genoma (Genoma &a,Genoma &b)> crossover,
                                 unsigned int best);
